Open URLs in new window
Adapted functionality according to #10 such that in can be chosen
whether a call to openUrl will create a new tab in the current window or
in a new one instead.

There are still some problems with the execution of content scripts
within the newly created tab. Execution works only sporadically and with
host permissions. ActiveTab permission seems to be without any effect.

Needs further investigation