Content script injection for #10
Content script runs in every tab now (configured in manifest) and sends
a message containing the links to be called to the background instead of
being injected in the single add-on tab only. 
The background script checks all received messages whether they are
originated from the correct tab for further proceeding.